Pull out an interface PersistentLockService
This is preparing to make Rocks and JDBC KVS's take an
NoOpPersistentLockService, since CAS is not supported on those KVS's.
