+++ atlasdb-impl-shared/src/test/java/com/palantir/atlasdb/sweep/CellsSweeperShould.java
@@ -46,6 +46,7 @@ import com.palantir.atlasdb.keyvalue.api.Namespace;
 import com.palantir.atlasdb.keyvalue.api.TableReference;
 import com.palantir.atlasdb.persistentlock.LockEntry;
 import com.palantir.atlasdb.persistentlock.PersistentLockService;
+import com.palantir.atlasdb.persistentlock.StandardPersistentLockService;
 import com.palantir.atlasdb.transaction.api.Transaction;
 
 public class CellsSweeperShould {
@@ -66,7 +67,7 @@ public class CellsSweeperShould {
 
     private final KeyValueService mockKvs = mock(KeyValueService.class);
     private final Follower mockFollower = mock(Follower.class);
-    private final PersistentLockService mockPls = mock(PersistentLockService.class);
+    private final PersistentLockService mockPls = mock(StandardPersistentLockService.class);
     private final LockEntry mockEntry = mock(LockEntry.class);
 
     private final CellsSweeper sweeper = new CellsSweeper(null, mockKvs, mockPls, 1, ImmutableList.of(mockFollower));
